requirements.txt
streamlit
openai
python-dotenv
azure-search-documents
azure-core